# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlasticWatch API",
    "description": "AI-GIS detection and prioritisation of likely plastic-waste hotspots. Reports show waste appears to be present; they do not establish who is responsible.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "description": "All tasks for an authority; only its own for a team.",
        "operationId": "list_tasks_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSummary"
                  },
                  "type": "array",
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "description": "Optimised route over VERIFIED hotspots. Each hotspot becomes cleanup_scheduled.",
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/stops/{stop_id}/arrive": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Arrive At Stop",
        "description": "Mark arrival \u2014 recorded only within ARRIVE_RADIUS_M (50 m) of the hotspot.",
        "operationId": "arrive_at_stop_tasks__task_id__stops__stop_id__arrive_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "stop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArriveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArriveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/stops/{stop_id}/after": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Upload After Photos",
        "description": "Two after-photos -> before/after record + SUGGESTED verdict (SPEC \u00a714).\n\nMoves the hotspot to cleanup_completed (the team's claim). Never resolves anything:\nan authority reviews the photos side by side. Retakes are allowed until reviewed.",
        "operationId": "upload_after_photos_tasks__task_id__stops__stop_id__after_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "stop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_after_photos_tasks__task_id__stops__stop_id__after_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeforeAfterRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/before-after": {
      "get": {
        "tags": [
          "before-after"
        ],
        "summary": "List Before After",
        "operationId": "list_before_after_before_after_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only records awaiting an authority's review.",
              "default": false,
              "title": "Pending"
            },
            "description": "Only records awaiting an authority's review."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeforeAfterRecord"
                  },
                  "title": "Response List Before After Before After Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/before-after/{ba_id}": {
      "get": {
        "tags": [
          "before-after"
        ],
        "summary": "Get Before After",
        "operationId": "get_before_after_before_after__ba_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ba_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ba Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeforeAfterRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/before-after/{ba_id}/review": {
      "post": {
        "tags": [
          "before-after"
        ],
        "summary": "Review Before After",
        "description": "Authority inspects before/after side by side and confirms or rejects.",
        "operationId": "review_before_after_before_after__ba_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ba_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ba Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArriveRequest": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "gps_accuracy_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Accuracy M"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "ArriveRequest",
        "description": "The team's current position, checked against the stop (SPEC \u00a78: within 50 m)."
      },
      "ArriveResponse": {
        "properties": {
          "stop": {
            "$ref": "#/components/schemas/TaskStop"
          },
          "distance_m": {
            "type": "number",
            "title": "Distance M"
          },
          "within_range": {
            "type": "boolean",
            "title": "Within Range"
          }
        },
        "type": "object",
        "required": [
          "stop",
          "distance_m",
          "within_range"
        ],
        "title": "ArriveResponse",
        "description": "Arrival is checked within 50 m of the stop (SPEC \u00a78)."
      },
...
      "BeforeAfterRecord": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "task_stop_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Stop Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "hotspot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotspot Id"
          },
          "hotspot_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HotspotStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "before_report_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Report Id"
          },
          "before_image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Image Path"
          },
          "before_annotated_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Annotated Path",
            "description": "Detector-annotated before photo (boxes), if written."
          },
          "after_image_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "After Image Paths",
            "default": []
          },
          "after_annotated_paths": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "After Annotated Paths",
            "description": "Detector-annotated after-photos, in after_image_paths order.",
            "default": []
          },
          "before_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Count"
          },
          "before_area": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Area"
          },
          "after_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Count"
          },
          "after_area": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Area"
          },
          "reduction_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reduction Ratio",
            "description": "1 - (after plastic area / before plastic area), using the worse photo."
          },
          "viewpoint_match": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewpoint Match",
            "description": "ORB match score against the before image."
          },
          "quality_flags": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality Flags",
            "default": {}
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict_note": {
            "type": "string",
            "title": "Verdict Note",
            "description": "CLAUDE.md \u00a72.5/\u00a72.6.",
            "default": "A verdict is a suggestion. Only an authority can resolve a hotspot."
          },
          "review_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "BeforeAfterRecord",
        "description": "SPEC \u00a714. The system computes a verdict but NEVER auto-resolves (CLAUDE.md \u00a72.5).\n\nAbsence of detections after cleanup is not proof of cleanliness (CLAUDE.md \u00a72.6)."
      },
...
      "Body_upload_after_photos_tasks__task_id__stops__stop_id__after_post": {
        "properties": {
          "wide": {
            "type": "string",
            "format": "binary",
            "title": "Wide",
            "description": "Wide after-photo."
          },
          "close": {
            "type": "string",
            "format": "binary",
            "title": "Close",
            "description": "Close-up after-photo."
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat",
            "description": "Team GPS at upload; else EXIF, else check-in."
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "wide",
          "close"
        ],
        "title": "Body_upload_after_photos_tasks__task_id__stops__stop_id__after_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotspotStatus": {
        "type": "string",
        "enum": [
          "ai_detected",
          "needs_verification",
          "verified",
          "cleanup_scheduled",
          "cleanup_completed",
          "resolved",
          "false_positive"
        ],
        "title": "HotspotStatus",
        "description": "SPEC \u00a713. Only an authority reaches verified / false_positive / resolved."
      },
...
      "ReviewRequest": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewRequest",
        "description": "A human authority action. Only confirm_resolved closes the hotspot."
      },
      "ReviewResponse": {
        "properties": {
          "before_after": {
            "$ref": "#/components/schemas/BeforeAfterRecord"
          },
          "hotspot_id": {
            "type": "integer",
            "title": "Hotspot Id"
          },
          "hotspot_status": {
            "$ref": "#/components/schemas/HotspotStatus"
          },
          "event": {
            "$ref": "#/components/schemas/HotspotEvent"
          }
        },
        "type": "object",
        "required": [
          "before_after",
          "hotspot_id",
          "hotspot_status",
          "event"
        ],
        "title": "ReviewResponse"
      },
...
      "TaskCreateRequest": {
        "properties": {
          "hotspot_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Hotspot Ids"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "depot": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Depot",
            "description": "[lon, lat] start/end point of the route."
          }
        },
        "type": "object",
        "required": [
          "hotspot_ids",
          "team_id",
          "depot"
        ],
        "title": "TaskCreateRequest",
        "description": "Tasks accept VERIFIED hotspots only (SPEC \u00a713) \u2014 a non-verified id is a 409."
      },
      "TaskDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "assigned_team": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team"
          },
          "assigned_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team Name"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "stop_count": {
            "type": "integer",
            "title": "Stop Count"
          },
          "route_distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Distance M"
          },
          "route_duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Duration S"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": false
          },
          "depot": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depot"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/TaskStop"
            },
            "type": "array",
            "title": "Stops"
          },
          "route_geojson": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Geojson",
            "description": "Decoded ORS geometry, or straight lines when the fallback ran."
          },
          "route_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ors",
                  "greedy"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "stop_count",
          "created_at",
          "stops"
        ],
        "title": "TaskDetail"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "planned",
          "in_progress",
          "done"
        ],
        "title": "TaskStatus",
        "description": "SPEC \u00a77 cleanup_tasks.status."
      },
...
      "TaskSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "assigned_team": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team"
          },
          "assigned_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Team Name"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "stop_count": {
            "type": "integer",
            "title": "Stop Count"
          },
          "route_distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Distance M"
          },
          "route_duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Duration S"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "stop_count",
          "created_at"
        ],
        "title": "TaskSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "likely_cleaned",
          "partial",
          "not_cleaned",
          "inconclusive"
        ],
        "title": "Verdict",
        "description": "SPEC \u00a714. A verdict is a suggestion \u2014 it never resolves anything by itself."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}